[Dispatch Creation] Merge bubbling of expand and extract
