+++ compiler/src/iree/compiler/DispatchCreation/BUILD.bazel
@@ -21,7 +21,6 @@ iree_compiler_cc_library(
     name = "DispatchCreation",
     srcs = [
         "BubbleUpExpandShapes.cpp",
-        "BubbleUpExtractSlices.cpp",
         "CloneProducersIntoDispatchRegions.cpp",
         "CollapseDimensions.cpp",
         "ConvertDispatchRegionsToWorkgroups.cpp",
